auto commint

+++ DOCKER/copy.sh
@@ -3,10 +3,10 @@ set -e
 
 # GIT ROOT
 PojectRoot="/home/nykl/Documents/WEB/Hradec_2020"
-#PojectRoot=""
 cd $PojectRoot
 
 
+# version inc
 source $PojectRoot/DOCKER/appVersion.txt
     #$priVersion = 
     #$sesVersion = 
@@ -19,16 +19,26 @@ sed -ir -e "s/\(terVersion=\).*/\1$newVersion/" "$PojectRoot/DOCKER/appVersion.t
 sed -ir -e "s/\(appVersion:\).*/\1$fullVersion,/" "$PojectRoot/HR/src/environments/environment.prod.ts"
 
 # python
-rm -f "$PojectRoot/DOCKER/apiService/requirements.txt"
+echo 'copy python'
+if [[ -f "$PojectRoot/DOCKER/apiService/requirements.txt" ]]; then
+    rm -f "$PojectRoot/DOCKER/apiService/requirements.txt"
+fi
 cp "$PojectRoot/FLUSK/requirements.txt" "$PojectRoot/DOCKER/apiService/requirements.txt"
 
-if [[ -f "$PojectRoot/FLUSK/app/__pycache__" ]]; then
+if [[ -d "$PojectRoot/FLUSK/app/__pycache__" ]]; then
     rm -r "$PojectRoot/FLUSK/app/__pycache__"
 fi
-
-rm -rf "$PojectRoot/DOCKER/apiService"
+if [[ -d "$PojectRoot/DOCKER/apiService/app" ]]; then
+   rm -rf "$PojectRoot/DOCKER/apiService/app"
+fi
 cp -r "$PojectRoot/FLUSK/app" "$PojectRoot/DOCKER/apiService/app"
 
+exit 0
+
 # html
+echo 'build angualr'
+if [[ -d "$PojectRoot/DOCKER/nginx/app" ]]; then
+   rm -rf "$PojectRoot/DOCKER/nginx/app/*"
+fi
 cd "$PojectRoot/HR"
 ng build --prod="true" --baseHref / --output-path="$PojectRoot/DOCKER/nginx/app"
